+++ test/unit/api-scheme.spec.js
@@ -15,9 +15,19 @@ const test = require('japa')
 const { ioc } = require('@adonisjs/fold')
 
 const { api: Api } = require('../../src/Schemes')
-const { lucid: LucidSerializer } = require('../../src/Serializers')
+const { lucid: LucidSerializer, database: DatabaseSerializer } = require('../../src/Serializers')
 const helpers = require('./helpers')
 const setup = require('./setup')
+
+const Encryption = {
+  encrypt (token) {
+    return `e${token}`
+  },
+
+  decrypt (token) {
+    return token.replace(/^e/, '')
+  }
+}
 
 test.group('Schemes - Api', (group) => {
   setup.databaseHook(group)
@@ -37,7 +47,7 @@ test.group('Schemes - Api', (group) => {
     const lucid = new LucidSerializer()
     lucid.setConfig(config)
 
-    const api = new Api()
+    const api = new Api(Encryption)
     api.setOptions(config, lucid)
 
     try {
@@ -63,7 +73,7 @@ test.group('Schemes - Api', (group) => {
 
     await User.create({ email: 'foo@bar.com', password: 'secret' })
 
-    const api = new Api()
+    const api = new Api(Encryption)
     api.setOptions(config, lucid)
 
     try {
@@ -87,7 +97,7 @@ test.group('Schemes - Api', (group) => {
 
     await User.create({ email: 'foo@bar.com', password: 'secret' })
 
-    const api = new Api()
+    const api = new Api(Encryption)
     api.setOptions(config, lucid)
     const validated = await api.validate('foo@bar.com', 'secret')
     assert.isTrue(validated)
@@ -107,7 +117,7 @@ test.group('Schemes - Api', (group) => {
 
     const user = await User.create({ email: 'foo@bar.com', password: 'secret' })
 
-    const api = new Api()
+    const api = new Api(Encryption)
     api.setOptions(config, lucid)
     const tokenPayload = await api.generate(user)
 
@@ -128,14 +138,14 @@ test.group('Schemes - Api', (group) => {
     lucid.setConfig(config)
 
     const user = await User.create({ email: 'foo@bar.com', password: 'secret' })
-    await user.tokens().create({ type: 'api_token', token: 22, is_revoked: false })
+    await user.tokens().create({ type: 'api_token', token: '22', is_revoked: false })
 
-    const api = new Api()
+    const api = new Api(Encryption)
     api.setOptions(config, lucid)
     api.setCtx({
       request: {
         header (key) {
-          return `Bearer 22`
+          return `Bearer e22`
         }
       }
     })
@@ -160,7 +170,7 @@ test.group('Schemes - Api', (group) => {
 
     await User.create({ email: 'foo@bar.com', password: 'secret' })
 
-    const api = new Api()
+    const api = new Api(Encryption)
     api.setOptions(config, lucid)
     api.setCtx({
       request: {
@@ -173,7 +183,7 @@ test.group('Schemes - Api', (group) => {
     try {
       await api.check()
     } catch ({ name, message }) {
-      assert.equal(message, 'E_INVALID_API_TOKEN: The api is invalid or missing')
+      assert.equal(message, 'E_INVALID_API_TOKEN: The api token is missing or invalid')
       assert.equal(name, 'InvalidApiToken')
     }
   })
@@ -193,7 +203,7 @@ test.group('Schemes - Api', (group) => {
     const user = await User.create({ email: 'foo@bar.com', password: 'secret' })
     await user.tokens().create({ type: 'api_token', token: 22, is_revoked: false })
 
-    const api = new Api()
+    const api = new Api(Encryption)
     api.setOptions(config, lucid)
     api.setCtx({
       request: {
@@ -222,7 +232,7 @@ test.group('Schemes - Api', (group) => {
     const user = await User.create({ email: 'foo@bar.com', password: 'secret' })
     await user.tokens().create({ type: 'api_token', token: 22, is_revoked: false })
 
-    const api = new Api()
+    const api = new Api(Encryption)
     api.setOptions(config, lucid)
     api.setCtx({
       request: {
@@ -237,5 +247,149 @@ test.group('Schemes - Api', (group) => {
 
     const isLogged = await api.check()
     assert.isTrue(isLogged)
+  })
+
+  test('throw exception when token is missing', async (assert) => {
+    assert.plan(3)
+    const User = helpers.getUserModel()
+
+    const config = {
+      model: User,
+      uid: 'email',
+      password: 'password'
+    }
+
+    const lucid = new LucidSerializer(ioc.use('Hash'))
+    lucid.setConfig(config)
+
+    const user = await User.create({ email: 'foo@bar.com', password: 'secret' })
+    await user.tokens().create({ type: 'api_token', token: 22, is_revoked: false })
+
+    const api = new Api(Encryption)
+    api.setOptions(config, lucid)
+    api.setCtx({
+      request: {
+        header () {
+          return null
+        },
+        input () {
+          return null
+        }
+      }
+    })
+
+    try {
+      await api.check()
+    } catch ({ name, message, status }) {
+      assert.equal(name, 'InvalidApiToken')
+      assert.equal(message, 'E_INVALID_API_TOKEN: The api token is missing or invalid')
+      assert.equal(status, 401)
+    }
+  })
+
+  test('return a list of tokens for a given user', async (assert) => {
+    const User = helpers.getUserModel()
+
+    const config = {
+      model: User,
+      uid: 'email',
+      password: 'password'
+    }
+
+    const lucid = new LucidSerializer(ioc.use('Hash'))
+    lucid.setConfig(config)
+
+    const user = await User.create({ email: 'foo@bar.com', password: 'secret' })
+    const api = new Api(Encryption)
+    api.setOptions(config, lucid)
+    const payload = await api.generate(user)
+    const tokensList = await api.listTokens(user)
+    assert.equal(tokensList.size(), 1)
+    assert.equal(tokensList.first().token, payload.token)
+  })
+
+  test('return fake response when no tokens exists', async (assert) => {
+    const User = helpers.getUserModel()
+
+    const config = {
+      model: User,
+      uid: 'email',
+      password: 'password'
+    }
+
+    const lucid = new LucidSerializer(ioc.use('Hash'))
+    lucid.setConfig(config)
+
+    const user = await User.create({ email: 'foo@bar.com', password: 'secret' })
+    const api = new Api(Encryption)
+    api.setOptions(config, lucid)
+    const tokensList = await api.listTokens(user)
+    assert.equal(tokensList.size(), 0)
+  })
+
+  test('return fake response when user is not defined', async (assert) => {
+    const User = helpers.getUserModel()
+
+    const config = {
+      model: User,
+      uid: 'email',
+      password: 'password'
+    }
+
+    const lucid = new LucidSerializer(ioc.use('Hash'))
+    lucid.setConfig(config)
+
+    const user = await User.create({ email: 'foo@bar.com', password: 'secret' })
+    const api = new Api(Encryption)
+    api.setOptions(config, lucid)
+    await api.generate(user)
+    const tokensList = await api.listTokens()
+    assert.equal(tokensList.size(), 0)
+  })
+
+  test('return a list of tokens via database serializer', async (assert) => {
+    const User = helpers.getUserModel()
+
+    const config = {
+      primaryKey: 'id',
+      table: 'users',
+      tokensTable: 'tokens',
+      uid: 'email',
+      foreignKey: 'user_id',
+      password: 'password'
+    }
+
+    const database = new DatabaseSerializer(ioc.use('Hash'))
+    database.setConfig(config)
+
+    const user = await User.create({ email: 'foo@bar.com', password: 'secret' })
+    const api = new Api(Encryption)
+    api.setOptions(config, database)
+    const payload = await api.generate(user)
+    const tokensList = await api.listTokens(user)
+    assert.lengthOf(tokensList, 1)
+    assert.equal(tokensList[0].token, payload.token)
+  })
+
+  test('generate token via user credentials', async (assert) => {
+    const User = helpers.getUserModel()
+
+    const config = {
+      model: User,
+      uid: 'email',
+      password: 'password'
+    }
+
+    const lucid = new LucidSerializer(ioc.use('Hash'))
+    lucid.setConfig(config)
+
+    await User.create({ email: 'foo@bar.com', password: 'secret' })
+
+    const api = new Api(Encryption)
+    api.setOptions(config, lucid)
+    const tokenPayload = await api.attempt('foo@bar.com', 'secret')
+
+    assert.isDefined(tokenPayload.token)
+    assert.equal(tokenPayload.type, 'bearer')
   })
 })